Add seasonality support to generate day, add keyword args
This commit actually adds support for weights/seasonality to generate-days,
which means the number of base timestamps per minute can be modified for
individual hours, thus generating trends or skew in the generated time series.

This commit also refactors the multi-arity options for generate-days into
keyword support, which makes it easier to supply options independently of each
other.